Fix remaining test import paths after modular restructuring
- Updated all test files outside src/lib/__tests__ to use new modular import paths
- Fixed performance, e2e, integration, and example test imports
- All 293 tests now passing successfully
- Complete modular restructuring is now fully functional
