+++ src/test/integration/recipe-manager.test.ts
@@ -1,4 +1,4 @@
-import { RecipeManager } from '../../lib/recipe-manager';
+import { RecipeManager } from '../../lib/core/services/recipe-manager';
 import { rootContainer, TOKENS, initializeServices, cleanupServices } from '../../lib/composition-root';
 import path from 'path';
